chore: refactor into packages
