+++ apps/api/src/repositories/auth/identity.ts
@@ -1,4 +1,4 @@
-import { identities } from "@db/schema/auth";
+import { identities } from "@shared/db/schema/auth";
 import type { DB } from "@middleware/db";
 import { HTTP_CODES } from "@src/config/http-codes";
 import { HTTPError } from "@src/lib/errors";
