+++ apps/api/src/repositories/auth/session.ts
@@ -1,4 +1,4 @@
-import { sessions } from "@db/schema/auth";
+import { sessions } from "@shared/db/schema/auth";
 import type { DB } from "@middleware/db";
 import { HTTP_CODES } from "@src/config/http-codes";
 import { HTTPError } from "@src/lib/errors";
