+++ apps/api/tsconfig.json
@@ -25,7 +25,7 @@
       "@services/*": ["src/services/*"],
       "@middleware/*": ["src/middleware/*"],
       "@handlers/*": ["src/handlers/*"],
-      "@db/*": ["src/db/*"],
+      "@db/*": ["../../packages/shared/src/db/*"],
       "@lib/*": ["src/lib/*"],
       "@integrations/*": ["src/integrations/*"],
       "@repositories/*": ["src/repositories/*"],
@@ -35,5 +35,5 @@
     }
   },
   "exclude": ["node_modules", "dist"],
-  "include": ["src/**/*", "tests/**/*"]
+  "include": ["src/**/*", "tests/**/*", "../../packages/shared/src/db"]
 }
