+++ packages/shared/scripts/populate-api-key.ts
