+++ packages/shared/src/db/schema/auth.ts
